Update CompressionTests.cs
:bug:
